NP-1241 Refactor structure of institution form
