+++ src/pages/admin/CustomerInstitutionAdminsForm.tsx
@@ -1,0 +1,9 @@
+import React, { FC } from 'react';
+
+const CustomerInstitutionAdminsForm: FC = () => {
+  // TODO: Fetch existing admins
+
+  return <>TODO: Admins form</>;
+};
+
+export default CustomerInstitutionAdminsForm;
